icu_locid_transform

Module fallback

Tools for locale fallback, enabling arbitrary input locales to be mapped into the nearest locale with data.

Modules

algorithms 🔒

Structs

LocaleFallbackConfig

Configuration settings for a particular fallback operation.

LocaleFallbackIterator

Iteration type for locale fallback operations.

LocaleFallbackIteratorInner 🔒

Inner iteration type. Does not own the item under fallback.

LocaleFallbackerBorrowed

Borrowed version of LocaleFallbacker.

LocaleFallbackerWithConfig

A LocaleFallbackerBorrowed with an associated LocaleFallbackConfig.

Enums

LocaleFallbackPriority

Hint for which subtag to prioritize during fallback.

LocaleFallbackSupplement

What additional data is required to load when performing fallback.