Struct icu_locid_transform::provider::AliasesV1

pub struct AliasesV1<'data> {
    pub language_variants: VarZeroVec<'data, StrStrPairVarULE>,
    pub sgn_region: ZeroMap<'data, TinyAsciiStr<3>, Language>,
    pub language_len2: ZeroMap<'data, TinyAsciiStr<2>, str>,
    pub language_len3: ZeroMap<'data, TinyAsciiStr<3>, str>,
    pub language: VarZeroVec<'data, StrStrPairVarULE>,
    pub script: ZeroMap<'data, TinyAsciiStr<4>, Script>,
    pub region_alpha: ZeroMap<'data, TinyAsciiStr<2>, Region>,
    pub region_num: ZeroMap<'data, TinyAsciiStr<3>, Region>,
    pub complex_region: ZeroMap<'data, TinyAsciiStr<3>, ZeroSlice<Region>>,
    pub variant: ZeroMap<'data, TinyAsciiStr<8>, Variant>,
    pub subdivision: ZeroMap<'data, TinyAsciiStr<7>, TinyAsciiStr<7>>,
}

This alias data is used for locale canonicalization. Each field defines a mapping from an old identifier to a new identifier, based upon the rules in from http://unicode.org/reports/tr35/#LocaleId_Canonicalization. The data is stored in sorted order, allowing for binary search to identify rules to apply. It is broken down into smaller vectors based upon some characteristic of the data, to help avoid unnecessary searches. For example, the sgn_region field contains aliases for sign language and region, so that it is not necessary to search the data unless the input is a sign language.

The algorithm in tr35 is not guaranteed to terminate on data other than what is currently in CLDR. For this reason, it is not a good idea to attempt to add or modify aliases for use in this structure.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Fields
language_variants: VarZeroVec<'data, StrStrPairVarULE>

[language(-variant)+\] -> [langid] This is not a map as itโs searched linearly according to the canonicalization rules.

sgn_region: ZeroMap<'data, TinyAsciiStr<3>, Language>

sgn-[region] -> [language]

language_len2: ZeroMap<'data, TinyAsciiStr<2>, str>

[language{2}] -> [langid]

language_len3: ZeroMap<'data, TinyAsciiStr<3>, str>

[language{3}] -> [langid]

language: VarZeroVec<'data, StrStrPairVarULE>

[langid] -> [langid] This is not a map as itโs searched linearly according to the canonicalization rules.

script: ZeroMap<'data, TinyAsciiStr<4>, Script>

[script] -> [script]

region_alpha: ZeroMap<'data, TinyAsciiStr<2>, Region>

[region{2}] -> [region]

region_num: ZeroMap<'data, TinyAsciiStr<3>, Region>

[region{3}] -> [region]

complex_region: ZeroMap<'data, TinyAsciiStr<3>, ZeroSlice<Region>>

[region] -> [region]+

variant: ZeroMap<'data, TinyAsciiStr<8>, Variant>

[variant] -> [variant]

subdivision: ZeroMap<'data, TinyAsciiStr<7>, TinyAsciiStr<7>>

[value{7}] -> [value{7}]