icu_locid_transform

Struct LocaleCanonicalizer

pub struct LocaleCanonicalizer { /* private fields */ }

Implements the algorithm defined in UTS #35: Annex C, LocaleId Canonicalization.

Examples

use icu::locid::Locale;
use icu::locid_transform::{LocaleCanonicalizer, TransformResult};

let lc = LocaleCanonicalizer::new();

let mut locale: Locale = "ja-Latn-fonipa-hepburn-heploc".parse().unwrap();
assert_eq!(lc.canonicalize(&mut locale), TransformResult::Modified);
assert_eq!(locale, "ja-Latn-alalc97-fonipa".parse().unwrap());

Implementations

impl LocaleCanonicalizer

pub const fn new() -> Self

A constructor which creates a LocaleCanonicalizer from compiled data.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_with_any_provider( provider: &(impl AnyProvider + ?Sized), ) -> Result<Self, LocaleTransformError>

A version of Self::new that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn try_new_unstable<P>(provider: &P) -> Result<Self, LocaleTransformError>
where P: DataProvider<AliasesV2Marker> + DataProvider<LikelySubtagsForLanguageV1Marker> + DataProvider<LikelySubtagsForScriptRegionV1Marker> + ?Sized,

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub const fn new_with_expander(expander: LocaleExpander) -> Self

Creates a LocaleCanonicalizer with a custom LocaleExpander and compiled data.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_with_expander_unstable<P>( provider: &P, expander: LocaleExpander, ) -> Result<Self, LocaleTransformError>
where P: DataProvider<AliasesV2Marker> + ?Sized,

A version of Self::new_with_expander that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn try_new_with_expander_with_any_provider( provider: &(impl AnyProvider + ?Sized), options: LocaleExpander, ) -> Result<Self, LocaleTransformError>

A version of Self::new_with_expander that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn canonicalize(&self, locale: &mut Locale) -> TransformResult

The canonicalize method potentially updates a passed in locale in place depending up the results of running the canonicalization algorithm from http://unicode.org/reports/tr35/#LocaleId_Canonicalization.

Some BCP47 canonicalization data is not part of the CLDR json package. Because of this, some canonicalizations are not performed, e.g. the canonicalization of und-u-ca-islamicc to und-u-ca-islamic-civil. This will be fixed in a future release once the missing data has been added to the CLDR json data. See: https://github.com/unicode-org/icu4x/issues/746

Examples

use icu::locid::Locale;
use icu::locid_transform::{LocaleCanonicalizer, TransformResult};

let lc = LocaleCanonicalizer::new();

let mut locale: Locale = "ja-Latn-fonipa-hepburn-heploc".parse().unwrap();
assert_eq!(lc.canonicalize(&mut locale), TransformResult::Modified);
assert_eq!(locale, "ja-Latn-alalc97-fonipa".parse().unwrap());