Module icu_locid_transform::provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

AliasesV1

This alias data is used for locale canonicalization. Each field defines a mapping from an old identifier to a new identifier, based upon the rules in from http://unicode.org/reports/tr35/#LocaleId_Canonicalization. The data is stored in sorted order, allowing for binary search to identify rules to apply. It is broken down into smaller vectors based upon some characteristic of the data, to help avoid unnecessary searches. For example, the sgn_region field contains aliases for sign language and region, so that it is not necessary to search the data unless the input is a sign language.

AliasesV2

This alias data is used for locale canonicalization. Each field defines a mapping from an old identifier to a new identifier, based upon the rules in from http://unicode.org/reports/tr35/#LocaleId_Canonicalization. The data is stored in sorted order, allowing for binary search to identify rules to apply. It is broken down into smaller vectors based upon some characteristic of the data, to help avoid unnecessary searches. For example, the sgn_region field contains aliases for sign language and region, so that it is not necessary to search the data unless the input is a sign language.

AliasesV1Marker

Marker type for AliasesV1: “locid_transform/aliases@1”

AliasesV2Marker

Marker type for AliasesV2: “locid_transform/aliases@2”

Baked

Baked data

CollationFallbackSupplementV1Marker

Marker type for LocaleFallbackSupplementV1: “fallback/supplement/co@1”

LanguageStrStrPair

A triplet of strings with a EncodeAsVarULE implementation.

LanguageStrStrPairVarULE

VarULE type for LanguageStrStrPair. See LanguageStrStrPair for documentation.

LikelySubtagsExtendedV1

This likely subtags data is used for full coverage of locales, including ones that don’t otherwise have data in the Common Locale Data Repository (CLDR).

LikelySubtagsExtendedV1Marker

Marker type for LikelySubtagsExtendedV1: “locid_transform/likelysubtags_ext@1”

LikelySubtagsForLanguageV1

This likely subtags data is used for the minimize and maximize operations. Each field defines a mapping from an old identifier to a new identifier, based upon the rules in https://www.unicode.org/reports/tr35/#Likely_Subtags.

LikelySubtagsForLanguageV1Marker

Marker type for LikelySubtagsForLanguageV1: “locid_transform/likelysubtags_l@1”

LikelySubtagsForScriptRegionV1

This likely subtags data is used for the minimize and maximize operations. Each field defines a mapping from an old identifier to a new identifier, based upon the rules in https://www.unicode.org/reports/tr35/#Likely_Subtags.

LikelySubtagsForScriptRegionV1Marker

Marker type for LikelySubtagsForScriptRegionV1: “locid_transform/likelysubtags_sr@1”

LikelySubtagsV1

This likely subtags data is used for the minimize and maximize operations. Each field defines a mapping from an old identifier to a new identifier, based upon the rules in https://www.unicode.org/reports/tr35/#Likely_Subtags.

LikelySubtagsV1Marker

Marker type for LikelySubtagsV1: “locid_transform/likelysubtags@1”

LocaleFallbackLikelySubtagsV1

Locale fallback rules derived from likely subtags data.

LocaleFallbackLikelySubtagsV1Marker

Marker type for LocaleFallbackLikelySubtagsV1: “fallback/likelysubtags@1”

LocaleFallbackParentsV1

Locale fallback rules derived from CLDR parent locales data.

LocaleFallbackParentsV1Marker

Marker type for LocaleFallbackParentsV1: “fallback/parents@1”

LocaleFallbackSupplementV1

Key-specific supplemental fallback data.

ScriptDirectionV1

This directionality data is used to determine the script directionality of a locale.

ScriptDirectionV1Marker

Marker type for ScriptDirectionV1: “locid_transform/script_dir@1”

StrStrPair

A pair of strings with a EncodeAsVarULE implementation.

StrStrPairVarULE

VarULE type for StrStrPair. See StrStrPair for documentation.

Constants

DEFAULT_REGION

ZZ is the most common region, so it is defaulted for data size savings.

DEFAULT_SCRIPT

Latn is the most common script, so it is defaulted for data size savings.