Module icu_locid_transform::provider::expander

Structs

LikelySubtagsExtendedV1

This likely subtags data is used for full coverage of locales, including ones that don’t otherwise have data in the Common Locale Data Repository (CLDR).

LikelySubtagsExtendedV1Marker

Marker type for LikelySubtagsExtendedV1: “locid_transform/likelysubtags_ext@1”

LikelySubtagsForLanguageV1

This likely subtags data is used for the minimize and maximize operations. Each field defines a mapping from an old identifier to a new identifier, based upon the rules in https://www.unicode.org/reports/tr35/#Likely_Subtags.

LikelySubtagsForLanguageV1Marker

Marker type for LikelySubtagsForLanguageV1: “locid_transform/likelysubtags_l@1”

LikelySubtagsForScriptRegionV1

This likely subtags data is used for the minimize and maximize operations. Each field defines a mapping from an old identifier to a new identifier, based upon the rules in https://www.unicode.org/reports/tr35/#Likely_Subtags.

LikelySubtagsForScriptRegionV1Marker

Marker type for LikelySubtagsForScriptRegionV1: “locid_transform/likelysubtags_sr@1”

LikelySubtagsV1

This likely subtags data is used for the minimize and maximize operations. Each field defines a mapping from an old identifier to a new identifier, based upon the rules in https://www.unicode.org/reports/tr35/#Likely_Subtags.

LikelySubtagsV1Marker

Marker type for LikelySubtagsV1: “locid_transform/likelysubtags@1”