Module icu_locid_transform::provider::fallback

Structs

CollationFallbackSupplementV1Marker

Marker type for LocaleFallbackSupplementV1: “fallback/supplement/co@1”

LocaleFallbackLikelySubtagsV1

Locale fallback rules derived from likely subtags data.

LocaleFallbackLikelySubtagsV1Marker

Marker type for LocaleFallbackLikelySubtagsV1: “fallback/likelysubtags@1”

LocaleFallbackParentsV1

Locale fallback rules derived from CLDR parent locales data.

LocaleFallbackParentsV1Marker

Marker type for LocaleFallbackParentsV1: “fallback/parents@1”

LocaleFallbackSupplementV1

Key-specific supplemental fallback data.

Constants

DEFAULT_REGION

ZZ is the most common region, so it is defaulted for data size savings.

DEFAULT_SCRIPT

Latn is the most common script, so it is defaulted for data size savings.